posix_timers: Remove dead task timer expiry caching
When reading a timer sample, posix_cpu_timer_get() and
posix_cpu_timer_schedule() both perform a caching of the timer expiry time
by converting its value from absolute to relative if the task has exited.

The reason for this caching is not clear though, it could be:

1) For performance reasons: no need to calculate the delta after the
   task has died, its cputime won't change anymore.  We can thus avoid
   some locking (sighand, tasklist_lock, rq->lock for task_delta_exec(),
   ...), and various operations to calculate the sample...

2) To keep the remaining delta for the timer available after the task
   has died.  When it gets reaped, its sighand disappears, so accessing
   the process wide cputime through tsk->signal is probably not safe.

Now, is the first reason really worth it?  I have no idea if it is a case
we really want to optimize.

Considering the second reason, we return a disarmed zero'ed timer when
tsk->sighand == NULL.  So if this is an assumed reason, it's broken.  And
this case only concern process wide timers that have their group leader
reaped.  The posix cpu timer shouldn't even be available anymore at that
time.  Unless the group leader changed since we called
posix_cpu_timer_create() after an exec?

Anyway for now I'm sending this as an RFC because there may well be subtle
things I left behind.

Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Cc: Stanislaw Gruszka <sgruszka@redhat.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>